fix(types): pyright strict on scorers — clean 70 errors, add blocking CI step
Cleans the 70 strict-mode errors in the scorers library and pins the 3
non-locked files (_common.py, cohen_malloy_classifier.py, fcff_yield.py)
behind a dedicated blocking CI step. Total project pyright drops from
1006 errors -> 936 (delta = 70 scorers strict).

opportunistic_form4.py is SHA256-locked per pre-reg ledger
(insider_pc_compound_2026_05_10) so it cannot carry the pragma overrides
+ type-narrowing the other 3 files use without invalidating the compound
audit verdict. The strict array drops the scorers/ dir entry in favour
of the 3 individual file paths; opportunistic_form4 stays in basic mode.

Per-file cleanup:
  _common.py — drop unnecessary cast() now that pandas-stubs ships
    correct overloads for Series.quantile/mean/std with scalar args
  cohen_malloy_classifier.py — type-narrow set.intersection via
    list[set[int]] dispatch on the first element
  fcff_yield.py — assert-narrow the 9 snapshot-dict locals after the
    any(v is None for v in ...) guard so pyright propagates non-None
    through the float() coercions
  pyrightconfig.json — strict array now enumerates the 3 non-locked
    scorer files explicitly; comment explains the pre-reg constraint

Pragmas at file head disable Unknown* family rules (pandas-stubs is
incomplete in strict mode for the method-call patterns these files use).
Load-bearing strict rules — reportArgumentType, reportReturnType,
reportAttributeAccessIssue, reportGeneralTypeIssues, reportCallIssue —
stay at error.

Dependencies: pandas-stubs>=3.0.0 added to dev group for narrowing.

CI: dedicated blocking step "Pyright strict (scorers — non-locked files)"
runs pyright on just the 3 file paths. The broader project-wide pyright
step stays continue-on-error against 936 remaining errors; follow-up PRs
will clean backtest (274) -> attribution (361) -> basic global (301).

Test: test_pyright_config.STRICT_ROOTS updated to the new file-level
entries, pinning the pre-reg-locked-file rationale.

+++ apps/alphalens-pipeline/alphalens_pipeline/scorers/_common.py
@@ -1,3 +1,4 @@
+# pyright: reportMissingTypeStubs=false, reportUnknownMemberType=false, reportUnknownArgumentType=false, reportUnknownVariableType=false, reportUnknownParameterType=false, reportUnknownLambdaType=false
 """Cross-sectional scoring helpers shared by Layer 2 screeners.
 
 These primitives are pure-function utilities used by multiple screener
@@ -31,8 +32,8 @@ def winsorize(
     valid = series.dropna()
     if valid.empty:
         return series
-    lo = float(valid.quantile(lower_pct))
-    hi = float(valid.quantile(upper_pct))
+    lo = valid.quantile(lower_pct)
+    hi = valid.quantile(upper_pct)
     return series.clip(lower=lo, upper=hi)
 
 
@@ -46,8 +47,8 @@ def rank_zscore(series: pd.Series) -> pd.Series:
     valid = series.dropna()
     if valid.empty:
         return pd.Series(np.nan, index=series.index)
-    mean = float(valid.mean())
-    std = float(valid.std(ddof=0))
+    mean = valid.mean()
+    std = valid.std(ddof=0)
     if not np.isfinite(std) or std == 0:
         return pd.Series(np.nan, index=series.index)
     return (series - mean) / std
